+++ alicompat/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## []
 
+- Added The Farlanders support
 - Initial release